refactor: restructure components and standardize naming conventions
- Move feature components into components/features/ (escrow, wallet, market, marketplace, landing)
- Move RouteWrapper to components/layout/
- Move icon.tsx to components/ui/Icon.tsx
- Rename footer.tsx -> Footer.tsx (PascalCase)
- Rename boarding-slide.tsx -> BoardingSlide.tsx (PascalCase)
- Rename onboarding-data.ts -> OnboardingData.ts (PascalCase)
- Add barrel index.ts exports to features/, layout/, and ui/ directories
- Update all import paths across app/ and components/ to reflect new structure
