+++ craft-nexus/components/features/landing/index.ts
@@ -1,0 +1,2 @@
+export { default as BrowseCategories } from "./BrowseCategories";
+export { default as CategoriesResult } from "./CategoriesResult";
