+++ craft-nexus/components/features/market/CourseCard.tsx
