+++ craft-nexus/components/features/market/index.ts
@@ -1,0 +1,3 @@
+export { default as CourseCard } from "./CourseCard";
+export type { CourseCardProps } from "./CourseCard";
+export { default as Sidebar } from "./Sidebar";
