+++ craft-nexus/components/features/marketplace/index.ts
@@ -1,0 +1,2 @@
+export { ProductCheckout, default } from "./ProductCheckout";
+export type { Product, ProductCheckoutProps } from "./ProductCheckout";
