+++ craft-nexus/components/ui/index.ts
@@ -1,0 +1,1 @@
+export { Facebook, Instagram } from "./Icon";
